pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1"
description = "-"
authors = ["Your Name <you@example.com>"]
license = "Mit License"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
django = "^4.2.6"
djangorestframework = "^3.14.0"
dj-database-url = "^2.1.0"
sqlalchemy = "^2.0.21"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.0.3"
pydantic = "^2.4.2"
pydantic-env = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
